# logger/logger_openapi.yaml
openapi: 3.0.0
info:
  title: Logger Service API
  version: 1.0.0
  description: API documentation for the Logger Service

servers:
  - url: http://localhost:5000
    description: Local development server

paths:
  /log:
    post:
      summary: Add a log entry
      operationId: addLog
      requestBody:
        description: Log entry data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                HOST_NAME:
                  type: string
                  example: "Server1"
                HOST_IP:
                  type: string
                  example: "192.168.1.1"
                SYSTEM_TYPE:
                  type: string
                  example: "Linux"
                LEVEL:
                  type: string
                  enum: ["INFO", "WARN", "DEBUG", "ERRO"]
                  example: "INFO"
                PROCESS_NAME:
                  type: string
                  example: "nginx"
                CONTENT:
                  type: string
                  example: "Service started successfully"
                LOG_TIME:
                  type: string
                  format: date-time
                  example: "2024-08-25T08:31:01"
      responses:
        '201':
          description: Log entry added successfully
        '400':
          description: Data missing
        '402':
          description: Data format wrong
        '502':
          description: Database connection failed
        '500':
          description: Internal server error

  /search:
    get:
      summary: Search log entries
      operationId: searchLogs
      parameters:
        - in: query
          name: host_name
          schema:
            type: string
          description: Host name to filter logs
        - in: query
          name: host_ip
          schema:
            type: string
          description: Host IP to filter logs
        - in: query
          name: system_type
          schema:
            type: string
          description: System type to filter logs
        - in: query
          name: level
          schema:
            type: array
            items:
              type: string
              enum: ["INFO", "WARN", "DEBUG", "ERRO"]
          description: Log level to filter logs
        - in: query
          name: log_time_start
          schema:
            type: string
            format: date-time
          description: Start time for log filtering
        - in: query
          name: log_time_end
          schema:
            type: string
            format: date-time
          description: End time for log filtering
      responses:
        '200':
          description: Logs retrieved successfully
        '500':
          description: Internal server error